Separate out styles and remove extra files
