--- DataFinder/src/client/Banner/Banner.scss
@@ -1,1 +1,0 @@
-@import '../DataFinder/DataFinder.scss'
