+++ DataFinder/src/client/DataFinder/components/Banner.scss
@@ -1,0 +1,26 @@
+
+.df-banner-wrapper {
+    margin-left: 15px;
+    margin-right: 15px;
+  }
+  
+  .df-banner-title {
+    float: left;
+    margin-right: 30px;
+  }
+  
+  .df-banner-titlebar {
+    border: 1px solid #aaaaaa;
+    background-color: white;
+    margin: 0;
+  }
+
+
+.df-banner-separator {
+    border-top: 1px solid #aaaaaa
+  }
+
+  .df-group-summary-counts {
+    margin-bottom: 5px;
+  }
+
