+++ DataFinder/src/client/DataFinder/components/Barplot.scss
@@ -1,0 +1,12 @@
+
+.datafinder-barplot {
+    overflow-y: auto;
+    // border-top: 1px solid #dddddd;
+    // border-bottom: 1px solid #dddddd
+  }
+  
+  .df-barplot-title {
+    text-align: center;
+    margin-bottom: 0;
+  }
+
