+++ DataFinder/src/client/DataFinder/components/DataFilters.scss
@@ -1,0 +1,114 @@
+// assay-data selector
+.df-assay-data-selector {
+  float: left;
+  margin: 10px;
+  border-left: 1px solid #bbbbbb;
+}
+
+.assay-data-dropdown {
+  & button {
+    width: max-content;
+  }
+  & i {
+    padding-left: 8px;
+  }
+
+  &.selected {
+    & button,
+    .open > .btn-default.dropdown-toggle:hover,
+    .open > .btn-default.dropdown-toggle:focus {
+      background-color: #238b45;
+      color: #fff;
+    }
+
+    & .open button {
+      background-color: #238b45;
+      color: #fff;
+    }
+  }
+}
+
+.df-data-filter-row {
+  display: flex;
+  width: 100%;
+  padding-bottom: 10px;
+  border-bottom: 1px dashed #bbb;
+  margin-left: 5px;
+  font-weight: normal;
+}
+.df-data-filter-row-element {
+  float: left;
+  padding: 3px;
+  text-transform: none;
+
+  & > span {
+    padding-top: 10px;
+    padding-bottom: 10px;
+    display: block;
+  }
+
+  & > button {
+    padding: 0;
+    margin-top: 6px;
+    border-radius: 5px;
+    border: 1px solid #bbb;
+    height: 25px;
+    width: 25px;
+    background-color: #238b45;
+    color: #fff;
+    border: 2px solid #f7f7f7;
+
+    &:hover {
+      border: 2px solid transparent;
+    }
+
+    &:disabled {
+      background-color: #a2caaf;
+      border: 2px solid #f7f7f7;
+    }
+
+    & > i {
+      margin-top: 4px;
+    }
+  }
+}
+
+.df-data-indicator {
+  display: inline-flex;
+  flex-wrap: wrap;
+  width: max-content;
+  &.short {
+    max-width: 200px;
+  }
+  &.long {
+    max-width: 300px;
+  }
+  &.longest {
+    max-width: 500px;
+  }
+}
+
+
+.assay-data-dropdown-label {
+    min-width: 8em;
+    text-align: left; 
+  }
+
+.df-andor-dropdown {
+  float: right;
+  margin-left: 15px;
+  min-width: auto;
+  padding: 0;
+
+  & button {
+    width: max-content;
+  }
+
+  & > ul {
+    min-width: auto;
+
+    & > li > a {
+      padding: 3px;
+    }
+  }
+}
