+++ DataFinder/src/client/DataFinder/components/FilterSelector.scss
@@ -1,0 +1,65 @@
+@import "./reusable/Dropdowns.scss";
+
+// Filter dropdowns
+.df-filter-dropdown {
+  max-height: 15em;
+  width: 5em;
+  display: table;
+
+  & .dropdown-menu {
+    width: 5em;
+  }
+
+  & button {
+    float: inline-start;
+    width: 125px;
+    background-image: none;
+  }
+
+  & > div.open {
+    & > .dropdown-menu {
+      display: block;
+      padding: 0;
+      /*left:100%;*/
+      /*top:10px;*/
+      height: auto;
+      max-height: 15em;
+      overflow-y: auto;
+      overflow-x: hidden;
+      white-space: normal;
+
+      & .checkbox {
+        padding-left: 5px;
+      }
+
+      & label:hover {
+        font-weight: bold;
+      }
+
+      & > li > a:hover {
+        background-color: #238b45;
+      }
+    }
+  }
+}
+
+.filter-dropdown-set {
+  float: left;
+  padding: 10px;
+  width: 15em;
+}
+
+.filter-menu {
+    padding: 10px;
+    /*left:100%;*/
+    /*top:10px;*/
+    height: auto;
+    // max-height: 15em;
+    overflow: auto;
+  }
+
+
+.filter-indicator-list {
+    display: table-row;
+  }
+
