+++ DataFinder/src/client/DataFinder/components/FilterSummary.scss
@@ -1,0 +1,74 @@
+
+.filterbar {
+    // border: 1px solid #aaaaaa;
+    // margin-left: 0;
+    // margin-right: 0;
+    // margin-top: 10px;
+    // margin-bottom: 10px;
+    // background-color: #fff;
+    margin: 0
+  }
+
+
+.filter-indicator {
+    width: auto;
+    float: left;
+    border-radius: 1em;
+    padding: 2px 10px;
+    margin: 2px;
+    white-space: normal !important;
+    word-wrap: break-word;
+    text-align: left;
+    display: inline-table;
+    margin-left: 15px;
+  
+    &.Study {
+      background: #54278f;
+      color: white;
+    }
+  
+    &.Subject {
+      background: #6baed6;
+      color: white;
+    }
+  
+    &.Data {
+      background: #238b45;
+      color: white;
+    }
+  
+  }
+  
+  .filter-deletor {
+    @extend .filter-indicator;
+    width: max-content;
+    max-width: 15em;
+    border: 2px solid #f7f7f7;
+    position: relative;
+    text-transform: none;
+    margin-left: 0;
+  
+    & > span {
+      font-weight: lighter;
+      // white-space: normal;
+      color: white;
+      text-align: left;
+      // word-wrap: break-word;
+      padding-right: 15px;
+    }
+  
+    &:hover {
+      border: 2px solid transparent //#e57282
+    }
+  
+    & > i {
+      position: absolute;
+      top: 0.5em;
+      right: 0.5em;
+    }
+  
+    &.Data {
+      max-width: max-content;
+    }
+  }
+
