+++ DataFinder/src/client/DataFinder/components/Loader.tsx
@@ -1,4 +1,5 @@
 import React from 'react'
+import './Loader.scss'
 
 export const Loader: React.FC = () => {
     return <div>
