+++ DataFinder/src/client/DataFinder/components/StudyCard.scss
@@ -1,0 +1,66 @@
+
+.study-card {
+    border: 2px solid #bcbcbc;
+    padding: 5px;
+    margin: 10px;
+    border-radius: 10px;
+    width: 280px;
+    display: inline-table;
+  }
+  
+  .study-label {
+    display: block;
+    margin-top: 0;
+    margin-bottom: 5px;
+    font-variant: small-caps;
+    font-size: 14px;
+    line-height: 1.4;
+    height: 30px;
+  }
+  
+  .study-checkbox {
+    display: block;
+    margin: 5px 0;
+    position: relative;
+    float: left;
+  }
+  
+  .study-id {
+    font-variant: small-caps;
+    font-size: 14px;
+    line-height: 1.4;
+  }
+  
+  .study-pi {
+    float: right;
+  }
+
+  hr {
+    margin-top: 5px;
+    margin-bottom: 2px;
+  }
+  
+  .study-title {
+    display: block;
+    text-align: left;
+    font-weight: bold;
+    font-size: 0.9em;
+    margin-bottom: 10px;
+    margin-top: 10px;
+    height: 4em;
+    overflow: auto;
+  }
+  
+  .progress {
+    margin-bottom: 0px;
+    height: 10px;
+  }
+  
+  .card-text {
+    font-size: 0.9em;
+    margin: 0 15px 0 15px;
+  }
+  
+  .study-properties {
+    height: 7em;
+  }
