--- DataFinder/src/client/DataFinder/components/d3/HeatmapSelector.d3.test.ts
@@ -1,9 +1,0 @@
-import { drawHeatmapSelector } from './HeatmapSelector.d3'
-
-describe("drawHeatmapSelector", () => {
-
-    document.body.innerHTML =
-    '<div class="heatmap2"></div>'
-    test.skip("", () => {})
-    // drawHeatmapSelector()
-})
